Change logic of taking data of current name of dictionaries their count of all and remain words

+++ Laword/Laword/CurrentDictionary.swift
@@ -9,6 +9,6 @@ import Foundation
 
 struct  CurrentDictionary {
     let name: String?
-    let countOfAllWords: Int?
-    let countOfRemainWords: Int?
+    let countOfAllWords: Int16?
+    let countOfRemainWords: Int16?
 }
